+++ pkgs/test_core/lib/scaffolding.dart
@@ -218,7 +218,6 @@ void group(Object? description, dynamic Function() body,
     Object? tags,
     Map<String, dynamic>? onPlatform,
     int? retry,
-    bool allowEmpty = false,
     @Deprecated('Debug only') bool solo = false}) {
   _declarer.group(description.toString(), body,
       testOn: testOn,
@@ -227,7 +226,6 @@ void group(Object? description, dynamic Function() body,
       tags: tags,
       onPlatform: onPlatform,
       retry: retry,
-      allowEmpty: allowEmpty,
       solo: solo);
 
   // Force dart2js not to inline this function. We need it to be separate from