+++ pkgs/test/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.24.8-wip
 
-* Consider empty `group` to be test failures.
+* Consider `group` with no test cases, but with another callback such as
+  `setUp`, to be test failures.
 
 ## 1.24.7