+++ pkgs/test_api/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.6.2-wip
 
-* Consider empty `group` to be test failures.
+* Consider `group` with no test cases, but with another callback such as
+  `setUp`, to be test failures.
 
 ## 0.6.1